feat(interactions): `ActOn::act_on()` returns `ActionType`
