+++ src/plugins/interactions/src/traits.rs
@@ -1,14 +1,19 @@
+use crate::components::blocker::Blocker;
 use bevy::prelude::Entity;
 use bevy_rapier3d::prelude::CollisionEvent;
 use common::components::ColliderRoot;
 
-use crate::components::blocker::Blocker;
-
-pub(crate) mod damage_health;
 pub(crate) mod rapier_context;
 
+#[derive(Debug, PartialEq, Clone, Copy)]
+pub enum ActionType {
+	Once,
+	OncePerTarget,
+	Always,
+}
+
 pub trait ActOn<TTarget> {
-	fn act_on(&mut self, target: &mut TTarget);
+	fn act_on(&mut self, target: &mut TTarget) -> ActionType;
 }
 
 pub trait FromCollisionEvent {
